Refactor cinematic names
