+++ ozaria/engine/cinematic/commands/AbstractCommand.js
