+++ ozaria/engine/cinematic/dialogsystem/helper.js
